feat: enhance cloud-init server with group user data handling and improve vendor data response
* refactor code to put less in citypes and more in local handlers
