+++ cmd/cloud-init-server/vendordata_handlers.go
@@ -1,6 +1,13 @@
 package main
 
-import "net/http"
+import (
+	"fmt"
+	"net/http"
+
+	"github.com/OpenCHAMI/cloud-init/internal/smdclient"
+	"github.com/go-chi/chi/v5"
+	"github.com/rs/zerolog/log"
+)
 
 // VendorDataHandler godoc
 // @Summary Get vendor data
@@ -8,12 +15,34 @@ import "net/http"
 // @Produce plain
 // @Success 200 {string} string
 // @Router /vendor-data [get]
-func VendorDataHandler(w http.ResponseWriter, r *http.Request) {
-	payload := `#template: jinja
-#include
-{% for group_name in vendor_data.groups.keys() %}
-https://{{ vendor_data.cloud_init_base_url }}/{{ group_name }}.yaml
-{% endfor %}
-`
-	w.Write([]byte(payload))
+func VendorDataHandler(smd smdclient.SMDClientInterface) http.HandlerFunc {
+	return func(w http.ResponseWriter, r *http.Request) {
+		var urlId string = chi.URLParam(r, "id")
+		var id string
+		var err error
+		// If this request includes an id, it can be interrpreted as an impersonation request
+		if urlId == "" {
+			ip := getActualRequestIP(r)
+			// Get the component information from the SMD client
+			id, err = smd.IDfromIP(ip)
+			if err != nil {
+				log.Print(err)
+				w.WriteHeader(http.StatusUnprocessableEntity)
+				return
+			} else {
+				log.Printf("xname %s with ip %s found\n", id, ip)
+			}
+		}
+		groups, err := smd.GroupMembership(id)
+		if err != nil {
+			// If the group information is not available, return an empty list
+			groups = []string{}
+		}
+
+		payload := "#includ\n"
+		for _, group_name := range groups {
+			payload += fmt.Sprintf("http://cloud-init:27777/cloud-init/%s.yaml\n", group_name)
+		}
+		w.Write([]byte(payload))
+	}
 }
